requirements.txt
google-generativeai==0.3.2
pymongo==4.6.1
pandas==2.1.4
python-dotenv==1.0.0
tabulate==0.9.0
requests==2.31.0 